feat: 重构六种STM轨迹类型并修复多项严重bug
- 六种STM轨迹统一为L=25mm/v=5m/s/f=200Hz：圆形(Circle)、正方形(Square)、等边三角形(Triangle)、Z形(Z-Shape)、阿基米德螺旋向外(SpiralOut)、阿基米德螺旋向内(SpiralIn)
- 新增阿基米德螺旋弧长参数化LUT，保证匀速扫描
- Z形从锯齿波振荡改为标准Z字折线
- 修复：USB中断中执行重负载(Update_Full_Waveform_Buffer+ADC轮询)，改为中断只缓存数据、主循环处理
- 修复：设备序列号char[12]截断为半个UID，扩展为char[25]
- 修复：CMD_SET_STIMULATION长度校验过松，改为按类型分别校验
- 修复：跨上下文共享变量缺少volatile
- 修复：pin_bit冗余计算、ADC重复采样、port_num_map越界、double热路径改float
- 清理死代码：Calculate_FPS、FPS、sysTickDelta、HAL_Delay_us

+++ Core/Src/custom_math.c
@@ -10,7 +10,7 @@ float Euler_Distance(const float From[3], const float To[3])
         float diff = From[i] - To[i];
         TempDistance += diff * diff; // Avoid calling pow() for exponent 2, use direct multiplication
     }
-    return (float)sqrt(TempDistance);
+    return sqrtf(TempDistance);
 }
 
 void Vector3Lerp(float Result[3], const float From[3], const float To[3], float t)
@@ -30,7 +30,7 @@ void Vector3Cross(float Result[3], const float A[3], const float B[3])
 
 void Vector3Normalize(float A[3])
 {
-    float length = (float)sqrt(A[0] * A[0] + A[1] * A[1] + A[2] * A[2]);
+    float length = sqrtf(A[0] * A[0] + A[1] * A[1] + A[2] * A[2]);
     if (length > 0.0f)
     {
         A[0] /= length;
